requirements.txt
torch==2.5.1
torchvision==0.20.1
open_clip_torch==2.30.0
datasets==3.2.0
huggingface-hub==0.27.1
numpy==2.0.2
scipy==1.13.1
scikit-learn==1.6.1
matplotlib==3.9.4
tqdm==4.67.1
pillow==11.1.0
PyYAML==6.0.2
requests==2.32.3

